requirements.txt
scrapy>=2.11
playwright>=1.40
beautifulsoup4>=4.12
pandas>=2.0
nltk>=3.8
anthropic>=0.25
APScheduler>=3.10
jinja2>=3.1
python-dotenv>=1.0
httpx>=0.27
sqlite-utils>=3.35
pytest>=7.4
pytest-asyncio>=0.23
pytrends>=4.9
google-api-python-client>=2.100
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.2
requests>=2.31
